requirements.txt
easy-torch==1.3.3
easydict
packaging
setproctitle
pandas
scikit-learn
tables
sympy
openpyxl
setuptools==59.5.0
numpy==1.24.4
tqdm==4.67.1
tensorboard==2.18.0
torch>=2.0.0
torch-geometric>=2.3.0
scipy>=1.10.0
captum>=0.6.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel
pyyaml
